change lib to teams and moved translation
