+++ tabs/package.json
@@ -30,7 +30,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "eject": "react-scripts eject",
-    "lint": "eslint --fix --max-warnings 0 src/*.jsx src/components/sample/*.jsx src/components/*.jsx src/components/sample/lib/*.js"
+    "lint": "eslint --fix --max-warnings 0 src/*.jsx src/components/sample/*.jsx src/components/*.jsx src/components/sample/**/*.js"
   },
   "eslintConfig": {
     "extends": [
