+++ tabs/src/components/sample/RSCList.jsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import { CLIENT_APP_ID, RSC_API_URL } from './TabConstants';
-import { useTeamsFx } from "../sample/lib/useTeamsFx";
+import { useTeamsFx } from "../sample/teams/useTeamsFx";
 import { Table } from '@fluentui/react-northstar';
 import { useTranslation } from 'react-i18next';
